Cleanup TestROM example.
